+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased] - 2018-XX-XX
+### Fix
+- Fix content scripts should not be inserted if they are already present.
+
 ## [2.0.7] - 2018-05-09
 ### Fix
 - Fix backward compatibility issue with legacy API.